+++ CHANGELOG.md
@@ -6,9 +6,8 @@ All notable changes to this project will be documented in this file.
 - Updated `winit` to `0.30`.
 - Fixed build issues on Windows and macOS.
 - Fixed compatibility with CMake 4.0+.
-- Removed the `glsl-to-spirv` dependency.
 - Propagated the `serde` feature to `notan_egui`.
-- Set `shaderc` as default (and only) shader transpiler.
+- Set `shaderc` as default shader transpiler.
 
 ## v0.13.0 - 29/03/2025